+++ integration_tests/models/metric_definitions/base_sum_metric.yml
@@ -4,7 +4,7 @@ metrics:
     model: ref('fact_orders')
     label: Order Total ($)
     timestamp: order_date
-    time_grains: [day, week, month, all_time]
+    time_grains: [day, week, month]
     calculation_method: sum
     expression: order_total
     dimensions:
@@ -15,7 +15,7 @@ metrics:
     model: ref('fact_orders_duplicate')
     label: Order Total ($)
     timestamp: order_date
-    time_grains: [day, week, month, all_time]
+    time_grains: [day, week, month]
     calculation_method: sum
     expression: order_total
     dimensions:
@@ -26,7 +26,7 @@ metrics:
     model: ref('fact_orders')
     label: Order Total ($)
     timestamp: order_date
-    time_grains: [day, week, month, all_time]
+    time_grains: [day, week, month]
     calculation_method: sum
     expression: order_total
     window:
@@ -40,7 +40,7 @@ metrics:
     model: ref('fact_orders')
     label: Order Total ($)
     timestamp: order_date
-    time_grains: [day, week, month, all_time] 
+    time_grains: [day, week, month] 
     calculation_method: sum
     expression: order_total
     dimensions:
