+++ integration_tests/models/metric_definitions/case_when_metric.yml
@@ -4,7 +4,7 @@ metrics:
     model: ref('fact_orders')
     label: Order Total ($)
     timestamp: order_date
-    time_grains: [day, week, month, all_time]
+    time_grains: [day, week, month]
     calculation_method: sum
     expression: case when had_discount = true then 1 else 0 end
     dimensions:
