+++ integration_tests/models/metric_testing_models/base_sum_metric.sql
@@ -1,7 +1,7 @@
 select *
 from 
 {{ metrics.calculate(
-    [metric('base_sum_metric'), metric('base_sum_metric_duplicate')],
+    metric('base_sum_metric'),
     secondary_calculations=[
         metrics.period_over_period(comparison_strategy="difference", interval=1, alias = "1mth")
     ]
