+++ tests/functional/calculation_methods/test_average.py
