+++ tests/functional/calculation_methods/test_count.py
