+++ tests/functional/calculation_methods/test_min.py
