+++ tests/functional/calculation_methods/test_sum.py
